requirements.txt
transformers==4.24.0
faiss-cpu==1.7.3
scikit-learn==1.2.2
requests==2.31.0
torch==2.0.0
dashscope
